Implement middleware for SearchInDocument
