+++ Resources/Public/JavaScript/PageView/SearchInDocument.js
@@ -230,7 +230,7 @@ $(document).ready(function() {
         $.post(
             "/",
             {
-                eID: "tx_dlf_search_in_document",
+                middleware: "dlf/search-in-document",
                 q: $( "input[id='tx-dlf-search-in-document-query']" ).val(),
                 uid: $( "input[id='tx-dlf-search-in-document-id']" ).val(),
                 pid: $( "input[id='tx-dlf-search-in-document-pid']" ).val(),
